为macOS12作准备
1. 更新第三方库到最新
2. 添加脚本上传提示页面
3. 为macOS12最准备
