+++ uPic/Extensions/NSImage+Extension.swift
@@ -1,14 +1,25 @@
 //
-//  NSImage+Extension.swift
+//  NSImageExtension.swift
 //  uPic
 //
-//  Created by Svend Jin on 2019/12/12.
+//  Created by 侯猛 on 2019/10/24.
 //  Copyright © 2019 Svend Jin. All rights reserved.
 //
 
 import Cocoa
+import Foundation
 
 extension NSImage {
+    func resizeImage(size: NSSize) -> NSImage {
+        let targetFrame = NSRect(x: 0, y: 0, width: size.width, height: size.height)
+        let sourceImageRep = self.bestRepresentation(for: targetFrame, context: nil, hints: nil)
+        let targetImage = NSImage(size: size)
+        targetImage.lockFocus()
+        sourceImageRep?.draw(in: targetFrame)
+        targetImage.unlockFocus()
+        return targetImage
+    }
+    
     var pngData: Data? {
         guard let rep = NSBitmapImageRep(
             bitmapDataPlanes: nil,
@@ -35,4 +46,8 @@ extension NSImage {
         }
         return data
     }
+}
+
+extension NSImage {
+
 }
