+++ uPic/Extensions/String+Extension.swift
