+++ uPic/Views/WelcomeView/en.lproj/Welcome.strings
@@ -1,6 +1,6 @@
 
-/* Class = "NSTextFieldCell"; title = "Share Extension"; ObjectID = "5PO-nX-1ts"; */
-"5PO-nX-1ts.title" = "Share Extension";
+/* Class = "NSTextFieldCell"; title = "Upload via Extension"; ObjectID = "5PO-nX-1ts"; */
+"5PO-nX-1ts.title" = "Upload via Extension";
 
 /* Class = "NSTextFieldCell"; title = "Next"; ObjectID = "EN9-1Z-Sqq"; */
 "EN9-1Z-Sqq.title" = "Next";
@@ -8,8 +8,8 @@
 /* Class = "NSTextFieldCell"; title = "A powerful and simple file uploader tool.\nUpload any files to any services you like."; ObjectID = "J8p-JL-LaG"; */
 "J8p-JL-LaG.title" = "A powerful and simple file uploader tool.\nUpload any files to any services you like.";
 
-/* Class = "NSTextFieldCell"; title = "Due to system security restrictions, uPic cannot access files outside the sandbox. \nYou will not be able to use right click upload, share upload, command line upload, etc. without authorization. \nNote: uPic does not collect any information from you."; ObjectID = "KKR-Bg-AzW"; */
-"KKR-Bg-AzW.title" = "Due to system security restrictions, uPic cannot access files outside the sandbox. \nYou will not be able to use right click upload, share upload, command line upload, etc. without authorization. \nNote: uPic does not collect any information from you.";
+/* Class = "NSTextFieldCell"; title = "Limited by macOS, uPic can not access files outside of sandbox.\nIf not granted permission, uPic may not be able to upload some files.\nPlease be aware that uPic DO NOT collect any of your informations."; ObjectID = "KKR-Bg-AzW"; */
+"KKR-Bg-AzW.title" = "Limited by macOS, uPic can not access files outside of sandbox.\nIf not granted permission, uPic may not be able to upload some files.\nPlease be aware that uPic DO NOT collect any of your informations.";
 
 /* Class = "NSTextFieldCell"; title = "Authorization"; ObjectID = "RwC-HL-AhF"; */
 "RwC-HL-AhF.title" = "Authorization";
@@ -44,14 +44,14 @@
 /* Class = "NSTextFieldCell"; title = "Welcome to"; ObjectID = "pk7-xU-bny"; */
 "pk7-xU-bny.title" = "Welcome to";
 
-/* Class = "NSTextFieldCell"; title = "Finder Extension"; ObjectID = "vGP-mK-Tjk"; */
-"vGP-mK-Tjk.title" = "Finder Extension";
+/* Class = "NSTextFieldCell"; title = "Upload via Script"; ObjectID = "vGP-mK-Tjk"; */
+"vGP-mK-Tjk.title" = "Upload via Script";
 
 /* Class = "NSTextFieldCell"; title = "GitHub, Imgur, SM.MS, Qiniu KODO, Upyun USS, Aliyun OSS, Tencent Cloud COS, Gitee, Amazon S3, Baidu Cloud BOS, Lsky Pro"; ObjectID = "ve2-Sd-1j7"; */
 "ve2-Sd-1j7.title" = "GitHub, Imgur, SM.MS, Qiniu KODO, Upyun USS, Aliyun OSS, Tencent Cloud COS, Gitee, Amazon S3, Baidu Cloud BOS, Lsky Pro";
 
-/* Class = "NSTextFieldCell"; title = "Click share button and select \"Upload via uPic\"."; ObjectID = "xlO-Fz-Nxq"; */
-"xlO-Fz-Nxq.title" = "Click share button and select \"Upload via uPic\".";
+/* Class = "NSTextFieldCell"; title = "Click share button and select \"Upload via uPic\".\nor\nRight click on the file and select \"Upload via uPic\"."; ObjectID = "xlO-Fz-Nxq"; */
+"xlO-Fz-Nxq.title" = "Click share button and select \"Upload via uPic\".\nor\nRight click on the file and select \"Upload via uPic\".";
 
-/* Class = "NSTextFieldCell"; title = "Right click on the file and select \"Upload via uPic\"."; ObjectID = "xz2-3l-Jef"; */
-"xz2-3l-Jef.title" = "Right click on the file and select \"Upload via uPic\".";
+/* Class = "NSTextFieldCell"; title = "You can upload any files using terminal srcipt or Apple Script.\nCheck tutorial for more info."; ObjectID = "xz2-3l-Jef"; */
+"xz2-3l-Jef.title" = "You can upload any files using terminal srcipt or Apple Script.\nCheck tutorial for more info.";
